+++ CHANGELOG.md
@@ -761,6 +761,8 @@
 - Content and layout improvements to the actual spend tab on Report
 - Content and layout improvements to the actual spend upload page
 
+- Fix: 'spending breakdown' download is no longer empty
+
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-63...HEAD
 [release-63]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-62...release-63
 [release-62]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-61...release-62